WIP fixing dict access
