+++ python3_ws/src/calibration/scripts/depthai_helpers/calibration_utils.py
@@ -5,6 +5,7 @@ import glob
 import os
 import shutil
 import numpy as np
+from scipy.spatial.transform import Rotation as R
 import re
 import time
 import json
@@ -21,24 +22,35 @@ def setPolygonCoordinates(height, width):
     slope = 150
 
     p_coordinates = [
-            [[margin,margin], [margin, height-margin], [width-margin, height-margin], [width-margin, margin]],
+        [[margin, margin], [margin, height-margin],
+            [width-margin, height-margin], [width-margin, margin]],
 
-            [[margin,0], [margin,height], [width//2, height-slope], [width//2, slope]],
-            [[horizontal_shift, 0], [horizontal_shift, height], [width//2 + horizontal_shift, height-slope], [width//2 + horizontal_shift, slope]],
-            [[horizontal_shift*2-margin, 0], [horizontal_shift*2-margin, height], [width//2 + horizontal_shift*2-margin, height-slope], [width//2 + horizontal_shift*2-margin, slope]],
+        [[margin, 0], [margin, height], [width//2, height-slope], [width//2, slope]],
+        [[horizontal_shift, 0], [horizontal_shift, height], [
+            width//2 + horizontal_shift, height-slope], [width//2 + horizontal_shift, slope]],
+        [[horizontal_shift*2-margin, 0], [horizontal_shift*2-margin, height], [width//2 +
+                                                                               horizontal_shift*2-margin, height-slope], [width//2 + horizontal_shift*2-margin, slope]],
 
-            [[width-margin, 0], [width-margin, height], [width//2, height-slope], [width//2, slope]],
-            [[width-horizontal_shift, 0], [width-horizontal_shift, height], [width//2-horizontal_shift, height-slope], [width//2-horizontal_shift, slope]],
-            [[width-horizontal_shift*2+margin, 0], [width-horizontal_shift*2+margin, height], [width//2-horizontal_shift*2+margin, height-slope], [width//2-horizontal_shift*2+margin, slope]],
+        [[width-margin, 0], [width-margin, height], [width//2, height-slope], [width//2, slope]],
+        [[width-horizontal_shift, 0], [width-horizontal_shift, height], [width //
+                                                                         2-horizontal_shift, height-slope], [width//2-horizontal_shift, slope]],
+        [[width-horizontal_shift*2+margin, 0], [width-horizontal_shift*2+margin, height], [width //
+                                                                                           2-horizontal_shift*2+margin, height-slope], [width//2-horizontal_shift*2+margin, slope]],
 
-            [[0,margin], [width, margin], [width-slope, height//2], [slope, height//2]],
-            [[0,vertical_shift], [width, vertical_shift], [width-slope, height//2+vertical_shift], [slope, height//2+vertical_shift]],
-            [[0,vertical_shift*2-margin], [width, vertical_shift*2-margin], [width-slope, height//2+vertical_shift*2-margin], [slope, height//2+vertical_shift*2-margin]],
+        [[0, margin], [width, margin], [
+            width-slope, height//2], [slope, height//2]],
+        [[0, vertical_shift], [width, vertical_shift], [width-slope,
+                                                        height//2+vertical_shift], [slope, height//2+vertical_shift]],
+        [[0, vertical_shift*2-margin], [width, vertical_shift*2-margin], [width-slope,
+                                                                          height//2+vertical_shift*2-margin], [slope, height//2+vertical_shift*2-margin]],
 
-            [[0,height-margin], [width, height-margin], [width-slope, height//2], [slope, height//2]],
-            [[0,height-vertical_shift], [width, height-vertical_shift], [width-slope, height//2-vertical_shift], [slope, height//2-vertical_shift]],
-            [[0,height-vertical_shift*2+margin], [width, height-vertical_shift*2+margin], [width-slope, height//2-vertical_shift*2+margin], [slope, height//2-vertical_shift*2+margin]]
-        ]
+        [[0, height-margin], [width, height-margin],
+         [width-slope, height//2], [slope, height//2]],
+        [[0, height-vertical_shift], [width, height-vertical_shift], [width -
+                                                                      slope, height//2-vertical_shift], [slope, height//2-vertical_shift]],
+        [[0, height-vertical_shift*2+margin], [width, height-vertical_shift*2+margin], [width -
+                                                                                        slope, height//2-vertical_shift*2+margin], [slope, height//2-vertical_shift*2+margin]]
+    ]
     return p_coordinates
 
 
@@ -96,64 +108,45 @@ class StereoCalibration(object):
 
         for camera in board_config.cameras.keys():
             cam_info = board_config.cameras[camera]
-            images_path = 
-        
+            images_path = filepath + '/' + cam_info.name
+            ret, intrinsics, dist_coeff, _, _, size = self.calibrate_intrinsics(images_path)
+            cam_info['intrinsics'] = intrinsics
+            cam_info['dist_coeff'] = dist_coeff
+            cam_info['size'] = size
+            cam_info['reprojection_error'] = ret
 
-        self.calib_data = [R1_fp32, R2_fp32, M1_fp32, M2_fp32, M3_fp32, R_fp32, T_fp32, R_rgb_fp32, T_rgb_fp32, d1_coeff_fp32, d2_coeff_fp32, d3_coeff_fp32]
-        
-        if 1:  # Print matrices, to compare with device data
-            np.set_printoptions(suppress=True, precision=6)
-            print("\nR1 (left)");  print(R1_fp32)
-            print("\nR2 (right)"); print(R2_fp32)
-            print("\nM1 (left)");  print(M1_fp32)
-            print("\nM2 (right)"); print(M2_fp32)
-            print("\nR");          print(R_fp32)
-            print("\nT");          print(T_fp32)
-            print("\nM3 (rgb)");   print(M3_fp32)
-            print("\nR (rgb)");    print(R_rgb_fp32)
-            print("\nT (rgb)");    print(T_rgb_fp32)
+        for camera in board_config.cameras.keys():
+            left_cam_info = board_config.cameras[camera]
+            if left_cam_info.has_key('extrinsics'):
+                if left_cam_info.extrinsics.has_key('to_cam'):
+                    left_cam = camera
+                    right_cam = left_cam_info.extrinsics.to_cam
+                    left_path = filepath + '/' + left_cam_info.name
 
-        if 0:  # Print computed homography, to compare with device data
-            np.set_printoptions(suppress=True, precision=6)
-            for res_height in [800, 720, 400]:
-                m1 = np.copy(M1_fp32)
-                m2 = np.copy(M2_fp32)
-                if res_height == 720:
-                    m1[1, 2] -= 40
-                    m2[1, 2] -= 40
-                if res_height == 400:
-                    m_scale = [[0.5,   0, 0],
-                               [0, 0.5, 0],
-                               [0,   0, 1]]
-                    m1 = np.matmul(m_scale, m1)
-                    m2 = np.matmul(m_scale, m2)
-                h1 = np.matmul(np.matmul(m2, R1_fp32), np.linalg.inv(m1))
-                h2 = np.matmul(np.matmul(m2, R2_fp32), np.linalg.inv(m2))
-                h1 = np.linalg.inv(h1)
-                h2 = np.linalg.inv(h2)
-                print('\nHomography H1, H2 for height =', res_height)
-                print(h1)
-                print()
-                print(h2)
+                    right_cam_info = board_config.cameras[left_cam_info.extrinsics.to_cam]
+                    right_path = filepath + '/' + right_cam_info.name
 
-        print("\tTook %i seconds to run image processing." %
-              (round(time.time() - start_time, 2)))
+                    specTranslation = left_cam_info.extrinsics.specTranslation
+                    rot = left_cam_info.extrinsics.rotation
 
-        # self.create_save_mesh()
+                    translation = [specTranslation.x, specTranslation.y, specTranslation.z]
+                    rotation = R.from_euler('xyz', [rot.r, rot.p, rot.y], degrees=True)
+                    
+                    extrinsics = self.calibrate_extrinsics(left_path, right_path, left_cam_info['intrinsics'], left_cam_info['dist_coeff'], right_cam_info['intrinsics'], right_cam_info['dist_coeff'], translation, rotation)
+                    if extrinsics[0] != -1:
+                        return extrinsics
 
-        epipolar_RRgb = None
-        epipolar_LR = None
-        rgb_reprojectErrors = None
-        if self.calibrate_rgb:
-            rgb_reprojectErrors = self.ret_rgb_scaled
-            if self.calibrate_LR:
-                epipolar_RRgb = self.test_epipolar_charuco_rgbr(filepath)
+                    if board_config.stereo_config.left_cam == left_cam and board_config.stereo_config.right_cam == right_cam:
+                        board_config.stereo_config.rectification_left = extrinsics[3]
+                        board_config.stereo_config.rectification_right = extrinsics[4]
+
+                    left_cam_info.extrinsics.rotation_matrix = extrinsics[1]
+                    left_cam_info.extrinsics.translation = extrinsics[2]
+                    left_cam_info.extrinsics.stereo_error = extrinsics[0]
+                    left_cam_info.extrinsics.epipolar_error = self.test_epipolar_charuco(left_path, right_path, left_cam_info.intrinsics, left_cam_info.dist_coeff, right_cam_info.intrinsics, right_cam_info.dist_coeff, extrinsics[3], extrinsics[4])
     
-        if self.calibrate_LR:
-            epipolar_LR = self.test_epipolar_charuco_lr(filepath)
-    
-        return rgb_reprojectErrors, epipolar_LR, epipolar_RRgb, self.calib_data
-        
+        return [1, board_config]
+
     def analyze_charuco(self, images, scale_req=False, req_resolution=(800, 1280)):
         """
         Charuco base pose estimation.
@@ -217,8 +210,8 @@ class StereoCalibration(object):
                                      winSize=(5, 5),
                                      zeroZone=(-1, -1),
                                      criteria=criteria)
-                    allCorners.append(res2[1]) # Charco chess corners
-                    allIds.append(res2[2]) # charuco chess corner id's
+                    allCorners.append(res2[1])  # Charco chess corners
+                    allIds.append(res2[2])  # charuco chess corner id's
                     all_marker_corners.append(marker_corners)
                     all_marker_ids.append(ids)
                     all_recovered.append(recoverd)
@@ -231,81 +224,82 @@ class StereoCalibration(object):
         imsize = gray.shape
         return allCorners, allIds, all_marker_corners, all_marker_ids, imsize, all_recovered
 
-    def calibrate_charuco3D(self, filepath):
+    def calibrate_intrinsics(self, image_files):
+        image_files = glob.glob(image_files + "/*")
+        image_files.sort()
+        assert len(
+            image_files) != 0, "ERROR: Images not read correctly, check directory"
+
+        allCorners, allIds, _, _, imsize, _ = self.analyze_charuco(image_files)
+        if self.camera_model == 'perspective':
+            return self.calibrate_camera_charuco(allCorners, allIds, imsize[::-1]), imsize[::-1] # (Height, width)
+        else:
+            return self.calibrate_fisheye(allCorners, allIds, imsize[::-1]), imsize[::-1] # (Height, width)
+
+    def calibrate_extrinsics(self, images_left, images_right, M_l, d_l, M_r, d_r, guess_translation, guess_rotation):
         self.objpoints = []  # 3d point in real world space
         self.imgpoints_l = []  # 2d points in image plane.
         self.imgpoints_r = []  # 2d points in image plane.
 
-        # calcorners_l = []  # 2d points in image
-        # calcorners_r = []  # 2d points in image
-        # calids_l = []  # ids found in image
-        # calids_r = []  # ids found in image
+        images_left = glob.glob(images_left + "/*")
+        images_right = glob.glob(images_right + "/*")
 
-        images_left = glob.glob(filepath + "/left/*")
-        images_right = glob.glob(filepath + "/right/*")
-        # images_rgb = glob.glob(filepath + "/rgb/*")
-        # print("Images left path------------------->")
-        # print(images_left)
         images_left.sort()
         images_right.sort()
-        # images_rgb.sort()
 
         assert len(
-            images_left) != 0, "ERROR: Images not read correctly, check directory"
+            images_left) != 0, "ERROR: Images not found, check directory"
         assert len(
-            images_right) != 0, "ERROR: Images not read correctly, check directory"
-        # assert len(
-        #     images_rgb) != 0, "ERROR: Images not read correctly, check directory"
+            images_right) != 0, "ERROR: Images not found, check directory"
 
+        scale = None
+        scale_req = False
+        frame_left_shape = cv2.imread(images_left[0]).shape
+        frame_right_shape = cv2.imread(images_right[0]).shape
+        scalable_res = frame_left_shape
+        scaled_res = frame_right_shape
+        if frame_right_shape[0] < frame_left_shape[0] and frame_right_shape[1] < frame_left_shape[1]:
+            scale_req = True
+            scale = frame_right_shape[1] / frame_left_shape[1]
+        elif frame_right_shape[0] > frame_left_shape[0] and frame_right_shape[1] > frame_left_shape[1]:
+            scale_req = True
+            scale = frame_left_shape[1] / frame_right_shape[1]
+            scalable_res = frame_right_shape
+            scaled_res = frame_left_shape
+
+        if scale_req:
+            scaled_height = scale * scalable_res[0]
+            diff = scaled_height - scaled_res[0]
+            # if scaled_height <  smaller_res[0]:
+            if diff < 0:
+                scaled_res = (scaled_height, scaled_res[1])
+
+        # scaled_res = (scaled_height, )
+        M_lp = self.scale_intrinsics(M_l, frame_left_shape, scaled_res)
+        M_rp = self.scale_intrinsics(M_r, frame_right_shape, scaled_res)
+        
         print("~~~~~~~~~~~ POSE ESTIMATION LEFT CAMERA ~~~~~~~~~~~~~")
-        allCorners_l, allIds_l, _, _, imsize, _ = self.analyze_charuco(
-            images_left)
+        allCorners_l, allIds_l, _, _, imsize_l, _ = self.analyze_charuco(
+            images_left, scale_req, scaled_res)
 
         print("~~~~~~~~~~~ POSE ESTIMATION RIGHT CAMERA ~~~~~~~~~~~~~")
-        allCorners_r, allIds_r, _, _, imsize, _ = self.analyze_charuco(
-            images_right)
-        self.img_shape = imsize[::-1]
+        allCorners_r, allIds_r, _, _, imsize_r, _ = self.analyze_charuco(
+            images_right, scale_req, scaled_res)
 
-        # self.img_shape_rgb = imsize_rgb[::-1]
-        if self.cameraModel == 'perspective':
-            ret_l, self.M1, self.d1, rvecs, tvecs = self.calibrate_camera_charuco(
-                allCorners_l, allIds_l, self.img_shape)
-            ret_r, self.M2, self.d2, rvecs, tvecs = self.calibrate_camera_charuco(
-                allCorners_r, allIds_r, self.img_shape)
-        else:
-            ret_l, self.M1, self.d1, rvecs, tvecs = self.calibrate_fisheye(allCorners_l, allIds_l, self.img_shape)
-            ret_r, self.M2, self.d2, rvecs, tvecs = self.calibrate_fisheye(allCorners_r, allIds_r, self.img_shape)
-        # self.fisheye_undistort_visualizaation(images_left, self.M1, self.d1, self.img_shape)
-        # self.fisheye_undistort_visualizaation(images_right, self.M2, self.d2, self.img_shape)
+        assert imsize_r != imsize_l, "Left and right resolution scaling is wrong"
 
+        return self.calibrate_stereo(
+            allCorners_l, allIds_l, allCorners_r, allIds_r, imsize_r, M_lp, d_l, M_rp, d_r)
 
-        print("~~~~~~~~~~~~~RMS error of left and right~~~~~~~~~~~~~~")
-        print(ret_l)
-        print(ret_r)
-        ret, self.M1, self.d1, self.M2, self.d2, self.R, self.T, E, F = self.calibrate_stereo(allCorners_l, allIds_l, allCorners_r, allIds_r, self.img_shape, self.M1, self.d1, self.M2, self.d2)
-
-        print("~~~~~~~~~~~~~RMS error of L-R stereo~~~~~~~~~~~~~~")
-        print(ret)
-
-        # TODO(sachin): Fix rectify for Fisheye
-        if self.cameraModel == 'perspective':
-            self.R1, self.R2, self.P1, self.P2, self.Q, validPixROI1, validPixROI2 = cv2.stereoRectify(
-                                                                                        self.M1,
-                                                                                        self.d1,
-                                                                                        self.M2,
-                                                                                        self.d2,
-                                                                                        self.img_shape, self.R, self.T)
-        else:
-            self.R1, self.R2, self.P1, self.P2, self.Q = cv2.fisheye.stereoRectify(self.M1,
-                self.d1,
-                self.M2,
-                self.d2,
-                self.img_shape, self.R, self.T)	
-
-        self.H1 = np.matmul(np.matmul(self.M2, self.R1),
-                            np.linalg.inv(self.M1))
-        self.H2 = np.matmul(np.matmul(self.M2, self.R2),
-                            np.linalg.inv(self.M2))
+    def scale_intrinsics(self, intrinsics, originalShape, destShape):
+        scale = destShape[1] / originalShape[1]
+        scale_mat = np.array([[scale, 0, 0], [0, scale, 0], [0, 0, 1]])
+        scaled_intrinsics = np.matmul(scale_mat, intrinsics)
+        print("Scaled height offset : {}".format((originalShape[0] * scale - destShape[0]) / 2))
+        print("Scaled width offset : {}".format((originalShape[1] * scale - destShape[1]) / 2))
+        scaled_intrinsics[1][2] -= (originalShape[0] * scale - destShape[0]) / 2;
+        scaled_intrinsics[0][2] -= (originalShape[1] * scale - destShape[1]) / 2;
+        return scaled_intrinsics
 
     def fisheye_undistort_visualizaation(self, img_list, K, D, img_size):
         for im in img_list:
@@ -314,23 +308,29 @@ class StereoCalibration(object):
             # h, w = img.shape[:2]
             if self.cameraModel == 'perspective':
                 map1, map2 = cv2.initUndistortRectifyMap(
-                                K, D, np.eye(3), K, img_size, cv2.CV_32FC1)
+                    K, D, np.eye(3), K, img_size, cv2.CV_32FC1)
             else:
-                map1, map2 = cv2.fisheye.initUndistortRectifyMap(K, D, np.eye(3), K, img_size, cv2.CV_32FC1)
-    
-            undistorted_img = cv2.remap(img, map1, map2, interpolation=cv2.INTER_LINEAR, borderMode=cv2.BORDER_CONSTANT)
+                map1, map2 = cv2.fisheye.initUndistortRectifyMap(
+                    K, D, np.eye(3), K, img_size, cv2.CV_32FC1)
+
+            undistorted_img = cv2.remap(
+                img, map1, map2, interpolation=cv2.INTER_LINEAR, borderMode=cv2.BORDER_CONSTANT)
             cv2.imshow("undistorted", undistorted_img)
             cv2.waitKey(0)
-            # cv2.destroyAllWindows()        
+            # cv2.destroyAllWindows()
 
-
-    def calibrate_camera_charuco(self, allCorners, allIds, imsize):
+    def calibrate_camera_charuco(self, allCorners, allIds, imsize, hfov):
         """
         Calibrates the camera using the dected corners.
         """
+        f = imsize[1] / (2 * np.tan(np.deg2rad(hfov/2)))
+        # TODO(sachin): Change the initialization to be initialized using the guess from fov
         print("CAMERA CALIBRATION")
         print(imsize)
-        if imsize[1] < 700:
+        cameraMatrixInit = np.array([[f,    0.0,      imsize[1]/2],
+                                     [0.0,     f,      imsize[0]/2],
+                                     [0.0,   0.0,        1.0]])
+        """ if imsize[1] < 700:
             cameraMatrixInit = np.array([[400.0,    0.0,      imsize[0]/2],
                                          [0.0,     400.0,  imsize[1]/2],
                                          [0.0,        0.0,        1.0]])
@@ -341,13 +341,13 @@ class StereoCalibration(object):
         else:
             cameraMatrixInit = np.array([[3819.8801,    0.0,     1912.8375],
                                          [0.0,     3819.8801, 1135.3433],
-                                         [0.0,        0.0,        1.]])
-        
+                                         [0.0,        0.0,        1.]]) """
+
         print("Camera Matrix initialization.............")
         print(cameraMatrixInit)
 
         distCoeffsInit = np.zeros((5, 1))
-        flags = (cv2.CALIB_USE_INTRINSIC_GUESS + 
+        flags = (cv2.CALIB_USE_INTRINSIC_GUESS +
                  cv2.CALIB_RATIONAL_MODEL + cv2.CALIB_FIX_ASPECT_RATIO)
     #     flags = (cv2.CALIB_RATIONAL_MODEL)
         (ret, camera_matrix, distortion_coefficients,
@@ -373,22 +373,21 @@ class StereoCalibration(object):
             for j in range(len(allIds[i])):
                 obj_pts_sub.append(one_pts[allIds[i][j]])
             obj_points.append(np.array(obj_pts_sub, dtype=np.float32))
-        
 
         cameraMatrixInit = np.array([[500,    0.0,      643.9126],
                                     [0.0,     500,  387.56018],
                                     [0.0,        0.0,        1.0]])
-        
+
         print("Camera Matrix initialization.............")
         print(cameraMatrixInit)
-        flags = cv2.fisheye.CALIB_RECOMPUTE_EXTRINSIC
+        flags = 0
         distCoeffsInit = np.zeros((4, 1))
         term_criteria = (cv2.TERM_CRITERIA_COUNT +
-                                    cv2.TERM_CRITERIA_EPS, 100, 1e-5)
-        
-        return cv2.fisheye.calibrate(obj_points, allCorners, imsize, cameraMatrixInit, distCoeffsInit, flags = flags, criteria = term_criteria)
-    
-    def calibrate_stereo(self, allCorners_l, allIds_l, allCorners_r, allIds_r, imsize, cameraMatrix_l, distCoeff_l, cameraMatrix_r, distCoeff_r):
+                         cv2.TERM_CRITERIA_EPS, 100, 1e-5)
+
+        return cv2.fisheye.calibrate(obj_points, allCorners, imsize, cameraMatrixInit, distCoeffsInit, flags=flags, criteria=term_criteria)
+
+    def calibrate_stereo(self, allCorners_l, allIds_l, allCorners_r, allIds_r, imsize, cameraMatrix_l, distCoeff_l, cameraMatrix_r, distCoeff_r, r_in, t_in):
         left_corners_sampled = []
         right_corners_sampled = []
         obj_pts = []
@@ -414,35 +413,45 @@ class StereoCalibration(object):
                 left_sub_corners.append(allCorners_l[i][j])
                 right_sub_corners.append(allCorners_r[i][idx])
                 obj_pts_sub.append(one_pts[allIds_l[i][j]])
-
-            obj_pts.append(np.array(obj_pts_sub, dtype=np.float32))
-            left_corners_sampled.append(
-                np.array(left_sub_corners, dtype=np.float32))
-            right_corners_sampled.append(
-                np.array(right_sub_corners, dtype=np.float32))
-
+            if len(left_sub_corners) > 3 and len(right_sub_corners) > 3:
+                obj_pts.append(np.array(obj_pts_sub, dtype=np.float32))
+                left_corners_sampled.append(
+                    np.array(left_sub_corners, dtype=np.float32))
+                right_corners_sampled.append(
+                    np.array(right_sub_corners, dtype=np.float32))
+            else:
+                return [-1, "Stereo Calib failed due to less common features"]
         stereocalib_criteria = (cv2.TERM_CRITERIA_COUNT +
                                 cv2.TERM_CRITERIA_EPS, 100, 1e-5)
 
         if self.cameraModel == 'perspective':
             flags = 0
-            flags |= cv2.CALIB_FIX_INTRINSIC # TODO(sACHIN): Try without intrinsic guess
+            flags |= cv2.CALIB_FIX_INTRINSIC
             flags |= cv2.CALIB_RATIONAL_MODEL
+            flags |= cv2.CALIB_USE_EXTRINSIC_GUESS
 
-            return cv2.stereoCalibrate(
+            ret, M1, d1, M2, d2, R, T, E, F = cv2.stereoCalibrate(
                 obj_pts, left_corners_sampled, right_corners_sampled,
                 cameraMatrix_l, distCoeff_l, cameraMatrix_r, distCoeff_r, imsize,
-                criteria=stereocalib_criteria, flags=flags)
-            
+                r_in, t_in, criteria=stereocalib_criteria, flags=flags)
+
+            R_l, R_r, P_l, P_r, Q, validPixROI1, validPixROI2 = cv2.stereoRectify(
+                cameraMatrix_l,
+                distCoeff_l,
+                cameraMatrix_r,
+                distCoeff_r,
+                imsize, R, T)
+            return [ret, R, T, R_l, R_r]
+
         elif self.cameraModel == 'fisheye':
-            print(len(obj_pts))
-            print('obj_pts')
-            # print(obj_pts) 
-            print(len(left_corners_sampled))
-            print('left_corners_sampled')
+            # print(len(obj_pts))
+            # print('obj_pts')
+            # print(obj_pts)
+            # print(len(left_corners_sampled))
+            # print('left_corners_sampled')
             # print(left_corners_sampled)
-            print(len(right_corners_sampled))
-            print('right_corners_sampled')
+            # print(len(right_corners_sampled))
+            # print('right_corners_sampled')
             # print(right_corners_sampled)
             for i in range(len(obj_pts)):
                 print('---------------------')
@@ -451,143 +460,130 @@ class StereoCalibration(object):
                 print(len(left_corners_sampled[i]))
                 print(len(right_corners_sampled[i]))
             flags = 0
-            # flags |= cv2.fisheye.CALIB_RECOMPUTE_EXTRINSIC # TODO(sACHIN): Try without intrinsic guess
-            return cv2.fisheye.stereoCalibrate(
+            # flags |= cv2.fisheye.CALIB_RECOMPUTE_EXTRINSIC # TODO(SACHIN): Try without intrinsic guess
+            ret, M1, d1, M2, d2, R, T, E, F = cv2.fisheye.stereoCalibrate(
                 obj_pts, left_corners_sampled, right_corners_sampled,
                 cameraMatrix_l, distCoeff_l, cameraMatrix_r, distCoeff_r, imsize,
                 flags=flags, criteria=stereocalib_criteria), None, None
-        
-    def rgb_calibrate(self, filepath):
-        images_rgb_pth = Path(filepath + "/rgb")
-        if not images_rgb_pth.exists():
-            raise RuntimeError("RGB dataset folder not found!! To skip rgb calibration use -drgb argument")
 
-        images_right = glob.glob(filepath + "/right/*")
-        images_rgb = glob.glob(filepath + "/rgb/*")
+            R_l, R_r, P_l, P_r, Q = cv2.fisheye.stereoRectify(
+                cameraMatrix_l,
+                distCoeff_l,
+                cameraMatrix_r,
+                distCoeff_r,
+                imsize, R, T)
+            return [ret, R, T, R_l, R_r]
 
-        images_right.sort()
-        images_rgb.sort()
-        if len(images_right) <= 0 or len(images_rgb) <= 0:
-            raise RuntimeError("Image Files not found")
+    def display_rectification(self, image_data_pairs):
+        print(
+            "Displaying Stereo Pair for visual inspection. Press the [ESC] key to exit.")
+        for image_data_pair in image_data_pairs:
+            img_concat = cv2.hconcat([image_data_pair[0], image_data_pair[1]])
+            img_concat = cv2.cvtColor(img_concat, cv2.COLOR_GRAY2RGB)
 
-        # frame_right_shape = cv2.imread(images_right[0]).shape
-        # frame_left_shape = cv2.imread(images_rgb[0]).shape
-        
-        allCorners_rgb_scaled, allIds_rgb_scaled, _, _, imsize_rgb_scaled, _ = self.analyze_charuco(
-            images_rgb, scale_req=True, req_resolution=(360, 640))
-        self.img_shape_rgb_scaled = imsize_rgb_scaled[::-1]
+            # draw epipolar lines for debug purposes
+            line_row = 0
+            while line_row < img_concat.shape[0]:
+                cv2.line(img_concat,
+                         (0, line_row), (img_concat.shape[1], line_row),
+                         (0, 255, 0), 1)
+                line_row += 30
 
-        self.ret_rgb_scaled, self.M3_scaled, self.d3_scaled, rvecs, tvecs = self.calibrate_camera_charuco(
-            allCorners_rgb_scaled, allIds_rgb_scaled, imsize_rgb_scaled[::-1])
+            # show image
+            cv2.imshow('Stereo Pair', img_concat)
+            k = cv2.waitKey(0)
+            if k == 27:  # Esc key to stop
+                break
 
-        print('vs. intrinisics computed after scaling the image --->')
-        scale = 1920/640
-        scale_mat = np.array([[scale, 0, 0], [0, scale, 0], [0, 0, 1]])
-        self.M3 = np.matmul(scale_mat, self.M3_scaled)
-        self.d3 = self.d3_scaled
-        
-        print("RGB callleded RMS at 720")
-        print(self.ret_rgb_scaled)
-        print('~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~')
-        # print(self.M3_scaled)
-        
-        if self.calibrate_LR:
-            allCorners_r_rgb, allIds_r_rgb, _, _, _, _ = self.analyze_charuco(
-                images_right, scale_req=True, req_resolution=(360, 640))
+                # os._exit(0)
+                # raise SystemExit()
 
-            # sampling common detected corners
-            rgb_scaled_rgb_corners_sampled = []
-            rgb_scaled_right_corners_sampled = []
-            rgb_scaled_obj_pts = []
-            rgb_pts = None
-            right_pts = None
-            one_pts = self.board.chessboardCorners
-            for i in range(len(allIds_rgb_scaled)):
-                rgb_sub_corners = []
-                right_sub_corners = []
-                obj_pts_sub = []
-                # if len(allIds_l[i]) < 70 or len(allIds_r[i]) < 70:
-                #     continue
-                for j in range(len(allIds_rgb_scaled[i])):
-                    idx = np.where(allIds_r_rgb[i] == allIds_rgb_scaled[i][j])
-                    if idx[0].size == 0:
-                        continue
-                    rgb_sub_corners.append(allCorners_rgb_scaled[i][j])
-                    right_sub_corners.append(allCorners_r_rgb[i][idx])
-                    obj_pts_sub.append(one_pts[allIds_rgb_scaled[i][j]])
+        cv2.destroyWindow('Stereo Pair')
 
-                rgb_scaled_obj_pts.append(
-                    np.array(obj_pts_sub, dtype=np.float32))
-                rgb_scaled_rgb_corners_sampled.append(
-                    np.array(rgb_sub_corners, dtype=np.float32))
-                rgb_scaled_right_corners_sampled.append(
-                    np.array(right_sub_corners, dtype=np.float32))
-                if rgb_pts is None:
-                    rgb_pts = np.array(rgb_sub_corners, dtype=np.float32)
-                    right_pts = np.array(right_sub_corners, dtype=np.float32)
-                else:
-                    np.vstack(
-                        (rgb_pts, np.array(rgb_sub_corners, dtype=np.float32)))
-                    np.vstack((right_pts, np.array(
-                        right_sub_corners, dtype=np.float32)))
+    def display_rectification(self, image_data_pairs):
+        print(
+            "Displaying Stereo Pair for visual inspection. Press the [ESC] key to exit.")
+        for image_data_pair in image_data_pairs:
+            img_concat = cv2.hconcat([image_data_pair[0], image_data_pair[1]])
+            img_concat = cv2.cvtColor(img_concat, cv2.COLOR_GRAY2RGB)
 
-            self.objpoints_rgb_r = rgb_scaled_obj_pts
-            self.imgpoints_rgb = rgb_scaled_rgb_corners_sampled
-            self.imgpoints_rgb_right = rgb_scaled_right_corners_sampled
-            
-            flags = 0
-            flags |= cv2.CALIB_FIX_INTRINSIC
-            flags |= cv2.CALIB_RATIONAL_MODEL
+            # draw epipolar lines for debug purposes
+            line_row = 0
+            while line_row < img_concat.shape[0]:
+                cv2.line(img_concat,
+                         (0, line_row), (img_concat.shape[1], line_row),
+                         (0, 255, 0), 1)
+                line_row += 30
 
-            stereocalib_criteria = (cv2.TERM_CRITERIA_COUNT +
-                                        cv2.TERM_CRITERIA_EPS, 100, 1e-5)
+            # show image
+            cv2.imshow('Stereo Pair', img_concat)
+            k = cv2.waitKey(0)
+            if k == 27:  # Esc key to stop
+                break
 
-            self.M2_rgb = np.copy(self.M2)
-            self.M2_rgb[1, 2] -= 60
-            self.d2_rgb = np.copy(self.d1)
-            
-            ret, _, _, _, _, self.R_rgb, self.T_rgb, E, F = cv2.stereoCalibrate(
-                self.objpoints_rgb_r, self.imgpoints_rgb, self.imgpoints_rgb_right,
-                self.M3_scaled, self.d3_scaled, self.M2_rgb, self.d2_rgb, self.img_shape_rgb_scaled,
-                criteria=stereocalib_criteria, flags=flags)
-            print("~~~~~~ Stereo calibration rgb-left RMS error ~~~~~~~~")
-            print(ret)
+                # os._exit(0)
+                # raise SystemExit()
 
-            # Rectification is only to test the epipolar
-            self.R1_rgb, self.R2_rgb, self.P1_rgb, self.P2_rgb, self.Q_rgb, validPixROI1, validPixROI2 = cv2.stereoRectify(
-                self.M3_scaled,
-                self.d3_scaled,
-                self.M2_rgb,
-                self.d2_rgb,
-                self.img_shape_rgb_scaled, self.R_rgb, self.T_rgb)
-        else:
-            self.R_rgb = np.zeros((3, 3), dtype=np.float32)
-            self.T_rgb = np.zeros(3, dtype=np.float32)
+        cv2.destroyWindow('Stereo Pair')
 
-    def test_epipolar_charuco_lr(self, dataset_dir):
-        images_left = glob.glob(dataset_dir + '/left/*.png')
-        images_right = glob.glob(dataset_dir + '/right/*.png')
+    def test_epipolar_charuco(self, left_img_pth, right_img_pth, M_l, d_l, M_r, d_r, r_l, r_r):
+        images_left = glob.glob(left_img_pth + '/*.png')
+        images_right = glob.glob(right_img_pth + '/*.png')
         images_left.sort()
         images_right.sort()
         print("<-----------------Left-Right camera Epiploar error---------------->")
         assert len(images_left) != 0, "ERROR: Images not read correctly"
         assert len(images_right) != 0, "ERROR: Images not read correctly"
+
+        scale = None
+        scale_req = False
+        frame_left_shape = cv2.imread(images_left[0]).shape
+        frame_right_shape = cv2.imread(images_right[0]).shape
+        scalable_res = frame_left_shape
+        scaled_res = frame_right_shape
+        if frame_right_shape[0] < frame_left_shape[0] and frame_right_shape[1] < frame_left_shape[1]:
+            scale_req = True
+            scale = frame_right_shape[1] / frame_left_shape[1]
+        elif frame_right_shape[0] > frame_left_shape[0] and frame_right_shape[1] > frame_left_shape[1]:
+            scale_req = True
+            scale = frame_left_shape[1] / frame_right_shape[1]
+            scalable_res = frame_right_shape
+            scaled_res = frame_left_shape
+
+        if scale_req:
+            scaled_height = scale * scalable_res[0]
+            diff = scaled_height - scaled_res[0]
+            # if scaled_height <  smaller_res[0]:
+            if diff < 0:
+                scaled_res = (scaled_height, scaled_res[1])
+        
+        M_lp = self.scale_intrinsics(M_l, frame_left_shape, scaled_res)
+        M_rp = self.scale_intrinsics(M_r, frame_right_shape, scaled_res)
+
+        print("~~~~~~~~~~~ POSE ESTIMATION LEFT CAMERA ~~~~~~~~~~~~~")
+        allCorners_l, allIds_l, _, _, imsize_l, _ = self.analyze_charuco(
+            images_left, scale_req, scaled_res)
+
+        print("~~~~~~~~~~~ POSE ESTIMATION RIGHT CAMERA ~~~~~~~~~~~~~")
+        allCorners_r, allIds_r, _, _, imsize_r, _ = self.analyze_charuco(
+            images_right, scale_req, scaled_res)
+
+        assert imsize_r != imsize_l, "Epipolar: Left and right resolution scaling is wrong"
+
         criteria = (cv2.TERM_CRITERIA_EPS +
                     cv2.TERM_CRITERIA_MAX_ITER, 100, 0.00001)
 
-        # if not use_homo:
         mapx_l, mapy_l = cv2.initUndistortRectifyMap(
-            self.M1, self.d1, self.R1, self.P1, self.img_shape, cv2.CV_32FC1)
+                                    M_l, d_l, r_l, M_r, imsize_r, cv2.CV_32FC1)
         mapx_r, mapy_r = cv2.initUndistortRectifyMap(
-            self.M2, self.d2, self.R2, self.P2, self.img_shape, cv2.CV_32FC1)
-
+                                    M_r, d_r, r_r, M_r, imsize_r, cv2.CV_32FC1)
         image_data_pairs = []
         for image_left, image_right in zip(images_left, images_right):
             # read images
             img_l = cv2.imread(image_left, 0)
             img_r = cv2.imread(image_right, 0)
-            # warp right image
 
+            # warp right image
             # img_l = cv2.warpPerspective(img_l, self.H1, img_l.shape[::-1],
             #                             cv2.INTER_CUBIC +
             #                             cv2.WARP_FILL_OUTLIERS +
@@ -624,7 +620,8 @@ class StereoCalibration(object):
                 marker_corners_l, ids_l, image_data_pair[0], self.board)
             res2_r = cv2.aruco.interpolateCornersCharuco(
                 marker_corners_r, ids_r, image_data_pair[1], self.board)
-            if res2_l[1] is not None and res2_l[2] is not None and len(res2_l[1]) > 3:
+
+            if res2_l[1] is not None and res2_r[2] is not None and len(res2_l[1]) > 3 and len(res2_r[1]) > 3:
 
                 cv2.cornerSubPix(image_data_pair[0], res2_l[1],
                                  winSize=(5, 5),
@@ -635,26 +632,27 @@ class StereoCalibration(object):
                                  zeroZone=(-1, -1),
                                  criteria=criteria)
 
-            # termination criteria
-            img_pth = Path(images_right[i])
-            name = img_pth.name
-            corners_l = []
-            corners_r = []
-            for j in range(len(res2_l[2])):
-                idx = np.where(res2_r[2] == res2_l[2][j])
-                if idx[0].size == 0:
-                    continue
-                corners_l.append(res2_l[1][j])
-                corners_r.append(res2_r[1][idx])
+                # termination criteria
+                img_pth = Path(images_right[i])
+                corners_l = []
+                corners_r = []
+                for j in range(len(res2_l[2])):
+                    idx = np.where(res2_r[2] == res2_l[2][j])
+                    if idx[0].size == 0:
+                        continue
+                    corners_l.append(res2_l[1][j])
+                    corners_r.append(res2_r[1][idx])
 
-            imgpoints_l.extend(corners_l)
-            imgpoints_r.extend(corners_r)
-            epi_error_sum = 0
-            for l_pt, r_pt in zip(corners_l, corners_r):
-                epi_error_sum += abs(l_pt[0][1] - r_pt[0][1])
-            
-            print("Average Epipolar Error per image on host in " + img_pth.name + " : " +
-                  str(epi_error_sum / len(corners_l)))
+                imgpoints_l.extend(corners_l)
+                imgpoints_r.extend(corners_r)
+                epi_error_sum = 0
+                for l_pt, r_pt in zip(corners_l, corners_r):
+                    epi_error_sum += abs(l_pt[0][1] - r_pt[0][1])
+                
+                print("Average Epipolar Error per image on host in " + img_pth.name + " : " +
+                    str(epi_error_sum / len(corners_l)))
+            else:
+                return -1
 
         epi_error_sum = 0
         for l_pt, r_pt in zip(imgpoints_l, imgpoints_r):
@@ -668,190 +666,17 @@ class StereoCalibration(object):
 
         return avg_epipolar
 
-    def test_epipolar_charuco_rgbr(self, dataset_dir):
-        images_rgb = glob.glob(dataset_dir + '/rgb/*.png')
-        images_right = glob.glob(dataset_dir + '/right/*.png')
-        images_rgb.sort()
-        images_right.sort()
 
-        print("<-----------------RGB-Right camera Epiploar error---------------->")
-        assert len(images_rgb) != 0, "ERROR: Images not read correctly"
-        assert len(images_right) != 0, "ERROR: Images not read correctly"
-        # criteria for marker detection/corner detections
-        criteria = (cv2.TERM_CRITERIA_EPS +
-                    cv2.TERM_CRITERIA_MAX_ITER, 100, 0.00001)
-        scale_width = 1280/self.img_shape_rgb_scaled[0]
-
-        mapx_rgb, mapy_rgb = cv2.initUndistortRectifyMap(
-            self.M3_scaled, self.d3_scaled, self.R1_rgb, self.M3_scaled, self.img_shape_rgb_scaled, cv2.CV_32FC1)
-        mapx_r, mapy_r = cv2.initUndistortRectifyMap(
-            self.M2_rgb, self.d2_rgb, self.R2_rgb, self.M3_scaled, self.img_shape_rgb_scaled, cv2.CV_32FC1)
-
-        # self.H1_rgb = np.matmul(np.matmul(self.M2, self.R1_rgb),
-        #                     np.linalg.inv(M_rgb))
-        # self.H2_r = np.matmul(np.matmul(self.M2, self.R2_rgb),
-        #                     np.linalg.inv(self.M2))
-
-        image_data_pairs = []
-        count = 0
-        for image_rgb, image_right in zip(images_rgb, images_right):
-            # read images
-            img_rgb = cv2.imread(image_rgb, 0)
-            img_r = cv2.imread(image_right, 0)
-            img_r = img_r[60: 420, :]
-
-            dest_res = (int(img_rgb.shape[1] * scale_width),
-                        int(img_rgb.shape[0] * scale_width))
-
-
-            # if img_rgb.shape[0] < 720:
-            #     raise RuntimeError("resizeed height of rgb is smaller than required. {0} < {1}".format(
-            #         img_rgb.shape[0], req_resolution[0]))
-            # del_height = (img_rgb.shape[0] - 720) // 2
-            # img_rgb = img_rgb[del_height: del_height + 720, :]
-            # self.parse_frame(img_rgb, "rectified_rgb_before",
-            #                  "rectified_"+str(count))
-            img_rgb = cv2.resize(img_rgb, (0, 0), fx=0.5, fy=0.5)
-            # warp right image
-
-            # img_rgb = cv2.warpPerspective(img_rgb, self.H1_rgb, img_rgb.shape[::-1],
-            #                             cv2.INTER_CUBIC +
-            #                             cv2.WARP_FILL_OUTLIERS +
-            #                             cv2.WARP_INVERSE_MAP)
-
-            # img_r = cv2.warpPerspective(img_r, self.H2_r, img_r.shape[::-1],
-            #                             cv2.INTER_CUBIC +
-            #                             cv2.WARP_FILL_OUTLIERS +
-            #                             cv2.WARP_INVERSE_MAP)
-
-            img_rgb = cv2.remap(img_rgb, mapx_rgb, mapy_rgb, cv2.INTER_LINEAR)
-            img_r = cv2.remap(img_r, mapx_r, mapy_r, cv2.INTER_LINEAR)
-            # self.parse_frame(img_rgb, "rectified_rgb", "rectified_"+str(count))
-            image_data_pairs.append((img_rgb, img_r))
-            count += 1
-
-        # compute metrics
-        imgpoints_r = []
-        imgpoints_l = []
-        for i, image_data_pair in enumerate(image_data_pairs):
-            #             gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
-            marker_corners_l, ids_l, rejectedImgPoints = cv2.aruco.detectMarkers(
-                image_data_pair[0], self.aruco_dictionary)
-            marker_corners_l, ids_l, _, _ = cv2.aruco.refineDetectedMarkers(image_data_pair[0], self.board,
-                                                                            marker_corners_l, ids_l,
-                                                                            rejectedCorners=rejectedImgPoints)
-
-            marker_corners_r, ids_r, rejectedImgPoints = cv2.aruco.detectMarkers(
-                image_data_pair[1], self.aruco_dictionary)
-            marker_corners_r, ids_r, _, _ = cv2.aruco.refineDetectedMarkers(image_data_pair[1], self.board,
-                                                                            marker_corners_r, ids_r,
-                                                                            rejectedCorners=rejectedImgPoints)
-
-            res2_l = cv2.aruco.interpolateCornersCharuco(
-                marker_corners_l, ids_l, image_data_pair[0], self.board)
-            res2_r = cv2.aruco.interpolateCornersCharuco(
-                marker_corners_r, ids_r, image_data_pair[1], self.board)
-            if res2_l[1] is not None and res2_l[2] is not None and len(res2_l[1]) > 3:
-
-                cv2.cornerSubPix(image_data_pair[0], res2_l[1],
-                                 winSize=(5, 5),
-                                 zeroZone=(-1, -1),
-                                 criteria=criteria)
-                cv2.cornerSubPix(image_data_pair[1], res2_r[1],
-                                 winSize=(5, 5),
-                                 zeroZone=(-1, -1),
-                                 criteria=criteria)
-
-            # termination criteria
-            corners_l = []
-            corners_r = []
-            for j in range(len(res2_l[2])):
-                idx = np.where(res2_r[2] == res2_l[2][j])
-                if idx[0].size == 0:
-                    continue
-                corners_l.append(res2_l[1][j])
-                corners_r.append(res2_r[1][idx])
-
-            imgpoints_l.extend(corners_l)
-            imgpoints_r.extend(corners_r)
-            epi_error_sum = 0
-            for l_pt, r_pt in zip(corners_l, corners_r):
-                epi_error_sum += abs(l_pt[0][1] - r_pt[0][1])
-            img_pth = Path(images_right[i])
-            # name = img_pth.name
-            print("Average Epipolar Error per image on host in " + img_pth.name + " : " +
-                  str(epi_error_sum / len(corners_l)))
-            
-        epi_error_sum = 0
-        for l_pt, r_pt in zip(imgpoints_l, imgpoints_r):
-            epi_error_sum += abs(l_pt[0][1] - r_pt[0][1])
-
-        avg_epipolar = epi_error_sum / len(imgpoints_r)
-        print("Average Epipolar Error of rgb_right: " + str(avg_epipolar))
-
-        if self.enable_rectification_disp:
-            self.display_rectification(image_data_pairs)
-
-        return avg_epipolar
-
-    def display_rectification(self, image_data_pairs):
-        print("Displaying Stereo Pair for visual inspection. Press the [ESC] key to exit.")
-        for image_data_pair in image_data_pairs:
-            img_concat = cv2.hconcat([image_data_pair[0], image_data_pair[1]])
-            img_concat = cv2.cvtColor(img_concat, cv2.COLOR_GRAY2RGB)
-
-            # draw epipolar lines for debug purposes
-            line_row = 0
-            while line_row < img_concat.shape[0]:
-                cv2.line(img_concat,
-                         (0, line_row), (img_concat.shape[1], line_row),
-                         (0, 255, 0), 1)
-                line_row += 30
-
-            # show image
-            cv2.imshow('Stereo Pair', img_concat)
-            k = cv2.waitKey(0)
-            if k == 27:  # Esc key to stop
-                break
-            
-                # os._exit(0)
-                # raise SystemExit()
-
-        cv2.destroyWindow('Stereo Pair')
-
-    def display_rectification(self, image_data_pairs):
-        print("Displaying Stereo Pair for visual inspection. Press the [ESC] key to exit.")
-        for image_data_pair in image_data_pairs:
-            img_concat = cv2.hconcat([image_data_pair[0], image_data_pair[1]])
-            img_concat = cv2.cvtColor(img_concat, cv2.COLOR_GRAY2RGB)
-
-            # draw epipolar lines for debug purposes
-            line_row = 0
-            while line_row < img_concat.shape[0]:
-                cv2.line(img_concat,
-                         (0, line_row), (img_concat.shape[1], line_row),
-                         (0, 255, 0), 1)
-                line_row += 30
-
-            # show image
-            cv2.imshow('Stereo Pair', img_concat)
-            k = cv2.waitKey(0)
-            if k == 27:  # Esc key to stop
-                break
-            
-                # os._exit(0)
-                # raise SystemExit()
-
-        cv2.destroyWindow('Stereo Pair')
-
-    def create_save_mesh(self): #, output_path):
+    def create_save_mesh(self):  # , output_path):
 
         curr_path = Path(__file__).parent.resolve()
         print("Mesh path")
         print(curr_path)
 
-        map_x_l, map_y_l = cv2.initUndistortRectifyMap(self.M1, self.d1, self.R1, self.M2, self.img_shape, cv2.CV_32FC1)
-        map_x_r, map_y_r = cv2.initUndistortRectifyMap(self.M2, self.d2, self.R2, self.M2, self.img_shape, cv2.CV_32FC1)
+        map_x_l, map_y_l = cv2.initUndistortRectifyMap(
+            self.M1, self.d1, self.R1, self.M2, self.img_shape, cv2.CV_32FC1)
+        map_x_r, map_y_r = cv2.initUndistortRectifyMap(
+            self.M2, self.d2, self.R2, self.M2, self.img_shape, cv2.CV_32FC1)
 
         """ 
         map_x_l_fp32 = map_x_l.astype(np.float32)
@@ -897,14 +722,14 @@ class StereoCalibration(object):
                             row_right.append(map_y_r[y, x])
                             row_right.append(map_x_r[y, x])
                 if (map_x_l.shape[1] % meshCellSize) % 2 != 0:
-                            row_left.append(0)
-                            row_left.append(0)
-                            row_right.append(0)
-                            row_right.append(0)
+                    row_left.append(0)
+                    row_left.append(0)
+                    row_right.append(0)
+                    row_right.append(0)
 
                 mesh_left.append(row_left)
-                mesh_right.append(row_right)    
-        
+                mesh_right.append(row_right)
+
         mesh_left = np.array(mesh_left)
         mesh_right = np.array(mesh_right)
         left_mesh_fpath = str(curr_path) + '/../resources/left_mesh.calib'
