+++ requirements.txt
@@ -4,6 +4,7 @@ requests==2.24.0
 argcomplete==1.12.1
 pygame==2.0.0
 rospkg==1.2.8
+access-dict-by-dot
 # depthai==0.3.0.0
 --extra-index-url https://artifacts.luxonis.com/artifactory/luxonis-python-snapshot-local/
 depthai==2.11.1.0.dev0+2b3964a5cf9635b0315ce53e19813369f2848280
